requirements.txt
python-telegram-bot==22.3
pandas==2.3.1
pyodbc==5.2.0
docxtpl==0.20.1
python-docx==1.2.0